Separated GOAPAgent from the core
This removes the need for passing in unused templates from the core of the GOAP planning. This is to enable the planner to be used with more than just moving agent objects.

Currently however GOAPBehaviour needs to go through some conversions to create the needed Queue of Agent Actions. Maybe there would be some way to avoid this.